Get the ER recorder working in helpers.
Test function can now be run with the optional "generateExpectedResults" paramter. If enabled, the ER_GEN directive will be active when running helpers. The only place this is currently relevant is in the the toMatchER function, which will (in ER_GEN mode), dump the actual result of the test as a recorded expected result.

This change also includes the json-dumper, which catches the dump from toMatchER and outputs it to a .json file.

In addition, toMatchER (in not-ER_GEN mode) will read from an expected .json file to compare the actual result, returning a pass or fail depending on what it finds.